refactor(ui): 调整客户端实现类的包结构并添加新类
