+++ TeamCode/src/main/java/org/betastudio/ftc/util/entry/SubmitAble.java
@@ -1,0 +1,5 @@
+package org.betastudio.ftc.util.entry;
+
+public interface SubmitAble {
+	void submit();
+}
