+++ TeamCode/src/main/java/org/firstinspires/ftc/opmodes/tests/ObjectionClientTest.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.opmodes.tests;
 import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.betastudio.ftc.ui.client.ObjectionClient;
+import org.betastudio.ftc.ui.client.implementation.ObjectionClient;
 import org.firstinspires.ftc.opmodes.teleops.TeleOpCore;
 
 @TeleOp(group = "9_Tests")
